raffle_server/requirements.txt
boto3==1.26.47
botocore==1.29.165
Flask==3.0.2
Flask_Cors==3.0.10
mysql-connector==2.2.9
cffi==1.15.1
six==1.16.0
pycparser==2.21